+++ yaml-tests/src/main/java/com/apple/foundationdb/relational/yamltests/command/QueryConfig.java
@@ -284,7 +284,7 @@ public abstract class QueryConfig {
                                             .setInsertReusedCount(actualPlannerMetrics.getLong(8)))
                             .build();
                     if (expectedPlannerMetricsInfo == null) {
-                        executionContext.putMetrics(blockName, currentQuery, lineNumber, actualInfo);
+                        executionContext.putMetrics(blockName, currentQuery, lineNumber, actualInfo, true);
                         logger.debug("⭐️ Successfully inserted new planner metrics at line {}", getLineNumber());
                     } else {
                         final var actualCountersAndTimers = actualInfo.getCountersAndTimers();
@@ -309,9 +309,9 @@ public abstract class QueryConfig {
                                 (actualCountersAndTimers.getInsertReusedCount() != expectedCountersAndTimers.getInsertReusedCount()
                                          && log("‼️ insertReusedCount differs; expected = " + expectedCountersAndTimers.getInsertReusedCount() +
                                         "; actual = " + actualCountersAndTimers.getInsertReusedCount(), getLineNumber())) || isDifferent;
+                        executionContext.putMetrics(blockName, currentQuery, lineNumber, actualInfo, isDifferent);
                         if (isDifferent) {
                             if (executionContext.testYamlRunnerOptions(YamlRunner.YamlRunnerOptions.CORRECT_STATS)) {
-                                executionContext.putMetrics(blockName, currentQuery, lineNumber, actualInfo);
                                 logger.debug("⭐️ Successfully updated planner metrics at line {}", getLineNumber());
                             } else {
                                 reportTestFailure("‼️ Planner metrics have changed for line " + getLineNumber());